+++ TombEngine/Objects/Generic/Switches/pulley_switch.h
@@ -6,5 +6,7 @@ struct ItemInfo;
 namespace TEN::Entities::Switches
 {
 	void InitializePulleySwitch(short itemNumber);
-	void PulleySwitchCollision(short itemNumber, ItemInfo* laraItem, CollisionInfo* coll);
+	void CollisionPulleySwitch(short itemNumber, ItemInfo* laraItem, CollisionInfo* coll);
+	void ControlPulleySwitch(short itemNumber);
+	bool TriggerPulley(short itemNumber, short timer);
 }