+++ CHANGELOG.md
@@ -93,6 +93,7 @@ TombEngine releases are located in this repository (alongside with Tomb Editor):
   - You must use this version: <https://github.com/TombEngine/Resources/raw/refs/heads/main/Wad2%20Objects/Traps/TR1_Thor%20Hammer.wad2>
 * Added TR3 Moving Laser.
 * Added TR4 Statue Plinth.
+* Added Breakable Wall. Can be used on ground or underwater.
 
 ### Lua API changes
 * Added diary module.